Refactor into a component
